requirements.txt
altair==5.3.0
annotated-types==0.7.0
appdirs==1.4.4
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
attrs==23.2.0
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.8.2
branca==0.7.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.1.1
cycler==0.12.1
dateutils==0.6.12
decorator==5.1.1
duckdb==1.0.0
entrypoints==0.4
et-xmlfile==1.1.0
executing==2.0.1
Faker==25.8.0
favicon==0.7.0
folium==0.16.0
fonttools==4.53.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
gw-dsl-parser==0.1.49
htbuilder==0.6.2
idna==3.7
importlib-metadata==7.1.0
importlib-resources==6.4.0
ipylab==1.0.0
ipython==8.12.3
ipywidgets==8.1.3
jedi==0.19.1
jinja2==3.1.4
joblib==1.4.2
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyterlab-widgets==3.0.11
kanaries-track==0.0.5
kiwisolver==1.4.5
lxml==5.2.2
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.7.5
matplotlib-inline==0.1.7
mdurl==0.1.2
monotonic==1.6
more-itertools==10.3.0
mysql-connector-python==8.4.0
numerize==0.12
numpy==1.24.4
openpyxl==3.1.3
packaging==24.0
pandas==2.0.3
parso==0.8.4
pickleshare==0.7.5
pillow==10.3.0
pkgutil-resolve-name==1.3.10
plotly==5.22.0
prometheus-client==0.20.0
prompt-toolkit==3.0.47
protobuf==4.25.3
psutil==6.0.0
pure-eval==0.2.2
pyarrow==16.1.0
pydantic==2.8.2
pydantic-core==2.20.1
pydeck==0.9.1
pygments==2.18.0
pygwalker==0.4.9.1
pymdown-extensions==10.8.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
scikit-learn==1.3.2
scipy==1.10.1
segment-analytics-python==2.2.3
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.31
sqlglot==25.5.1
st-annotated-text==4.0.1
st-theme==1.2.3
stack-data==0.6.3
streamlit==1.36.0
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.2
streamlit-embedcode==0.1.2
streamlit-extras==0.4.3
streamlit-faker==0.0.3
streamlit-folium==0.20.0
streamlit-image-coordinates==0.1.9
streamlit-keyup==0.2.4
streamlit-option-menu==0.3.13
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.3.0
threadpoolctl==3.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4
traitlets==5.14.3
types-python-dateutil==2.9.0.20240316
typing-extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
validators==0.28.3
wasmtime==22.0.0
watchdog==4.0.1
wcwidth==0.2.13
widgetsnbextension==4.0.11
xgboost==2.0.3
xlrd==2.0.1
xyzservices==2024.6.0
zipp==3.19.2
